pyproject.toml
[project]
name = "ormvis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2>=2.9.12",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
    "ruff>=0.15.11",
    "sqlalchemy>=2.0.49",
    "ty>=0.0.32",
]
